plugins.v2/p115offlinebridge/requirements.txt
grpcio~=1.78.0
protobuf~=5.29.6
